Version 0.38b --------------------------------------     + Fixed the z-coordinate Inversion when exporting/importing a Mesh     + Added a total Face/Vertex Count     + Vertex count will now Display the Number of really used Vertices (not the Number of referenced Vertices)

+++ trunk/fullsimpe/SimPe Packages/HeaderData.cs
@@ -114,7 +114,7 @@ namespace SimPe.Packages
 		{
 			get 
 			{
-				return ((long)MajorVersion << 0x20) | (long)MinorVersion;
+				return (long)((ulong)MajorVersion << 0x20) | (uint)MinorVersion;
 			}
 		}
